django/django_blog_fixed/requirements.txt
Django>=5.1,<5.2

dj-database-url>=2.2,<3.0
python-dotenv>=1.0,<2.0
whitenoise>=6.7,<7.0

# PostgreSQL driver (choose one)
psycopg[binary]>=3.2,<4.0

gunicorn>=22.0,<23.0
